dashboard/chatbot/requirements.txt
pandas
chromadb
sentence-transformers
streamlit
groq
python-dotenv